requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-whois==0.8.0
dnspython==2.4.2
aiohttp==3.9.1
asyncio==3.4.3
opencv-python==4.8.1.78
numpy==1.24.3
pyautogui==0.9.54
keyboard==0.13.5
selenium==4.15.2
python-nmap==0.7.1
pyaudio==0.2.13
jinja2==3.1.2
pdfkit==1.0.0
markdown==3.5.1
pywinauto==0.6.8
pywin32==306
protobuf==4.21.12
typing-extensions==4.5.0
pillow==10.0.0
moviepy==1.0.3
python-docx==0.8.11
xlsxwriter==3.1.2
psutil==5.9.8
huggingface-hub==0.19.4
PyJWT==2.8.0
scikit-learn==1.4.0 